pyproject.toml
[project]
name = "lithox"
version = "0.1.0"
description = "High performance jax-based photolithography simulation."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Thomas Hirtz", email = "thomas.hirtz@pm.me" }
]
license = "BSD-3-Clause"
license-files = ["LICENSE.md"]

dependencies = [
  "jax",
  "numpy",
  "equinox",
  "Pillow",
  "chex",
  "torch",
  "matplotlib",
  "requests",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["lithox*"]

[tool.setuptools.package-data]
"lithox" = [
  "kernels/*.npy",
  "scales/*.npy"
]